User import/upload refactoring
Identified a major bug. Seems using start_new_thread is evil. When
another request comes in while a job is running, it kills (?) the old
job. I'll need to look into this and make sure that's the case. I don't
think this happens while browsing the site, certainly not when
refreshing the games page.